split probe method out; split out structs into separate file
